pyproject.toml
[project]
name = "api-proxy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "browserbase>=1.3.0",
    "celery>=5.5.2",
    "fastapi-utilities>=0.3.1",
    "hypercorn>=0.17.3",
    "langchain[openai]>=0.3.25",
    "langfuse>=2.60.4",
    "langgraph>=0.4.3",
    "playwright>=1.52.0",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "redis>=6.1.0",
    "typing-inspect>=0.9.0",
]

[dependency-groups]
dev = [
    "notebook>=7.4.2",
]
